refactor: extract pnp.rs and keypoint_grid.rs, simplify map_projection estimator
- Extract PnP solving and reprojection inlier counting into estimation/pnp.rs
- Move KeypointGrid into map_projection/keypoint_grid.rs submodule
- Move frustum visibility query from estimator to Map::map_points_in_frustum
- Remove TrackAttempt enum, FrameMatchResult struct, refine_pose and try_track
  indirections — inline their logic directly in estimate_pose
- Replace hardcoded consts with PnpConfig fields (min_inliers_early)
